+++ resolwe/flow/executors/prepare.py
@@ -49,9 +49,9 @@ class BaseFlowExecutorPreparer:
         """
         pass
 
-    def get_tools_paths(self):
+    def get_tools_paths(self, from_applications=False):
         """Get tools' paths."""
-        if settings.DEBUG or is_testing():
+        if settings.DEBUG or is_testing() or from_applications:
             return list(get_apps_tools().values())
 
         else: