+++ docs/CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+===========
+
+Added
+-----
+- Add health checks support for deploy in Kubernetes
+
 ===================
 29.2.0 - 2021-01-17
 ===================