+++ docs/CHANGELOG.rst
@@ -13,6 +13,12 @@ Unreleased
 Added
 -----
 - Add health checks support for deploy in Kubernetes
+- Add ``collecttools_kubernetes`` management command
+
+Changed
+-------
+- Do not prepare tools configmaps in Kubernetes workload connector
+
 
 ===================
 29.2.0 - 2021-01-17