+++ docs/CHANGELOG.rst
@@ -14,6 +14,8 @@ Added
 -----
 - Add health checks support for deploy in Kubernetes
 - Add ``collecttools_kubernetes`` management command
+- Add ``COMMUNICATION_CONTAINER_LISTENER_CONNECTION`` to separate settings for
+  listener (where to bind to) and containers (where to connect to)
 
 Changed
 -------