+++ docs/CHANGELOG.rst
@@ -22,6 +22,7 @@ Added
 Changed
 -------
 - Do not prepare tools configmaps in Kubernetes workload connector
+- Enable Docker containers to connect to the custom network
 
 
 ===================