+++ docs/CHANGELOG.rst
@@ -24,6 +24,7 @@ Changed
 -------
 - Do not prepare tools configmaps in Kubernetes workload connector
 - Enable Docker containers to connect to the custom network
+- Auto-delete completed jobs in Kubernetes after 5 minutes
 
 
 ===================