+++ docs/CHANGELOG.rst
@@ -7,6 +7,15 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
 ===================
+29.2.1 - 2021-02-14
+===================
+
+Added
+-----
+- Add MD5 checksum to ``DataBrowseView`` view
+
+
+===================
 29.2.0 - 2021-01-17
 ===================